+++ bigmler/defaults.py
@@ -155,7 +155,8 @@ FLAGS = {
         {'flag': 'max_nodes', 'type': 'integer'},
         {'flag': 'min_nodes', 'type': 'integer'},
         {'flag': 'nodes_step', 'type': 'integer'},
-        {'flag': 'exclude_features', 'type': 'string'}],
+        {'flag': 'exclude_features', 'type': 'string'},
+        {'flag': 'optimize_category', 'type': 'string'},],
     'BigMLer cluster': [
         {'flag': 'cluster_fields', 'type': 'string'},
         {'flag': 'cluster', 'type': 'string'},